Moved tracking to its own file and function
