+++ lib/run.js
@@ -3,7 +3,6 @@
 const EE = require('events').EventEmitter
 const URL = require('url')
 const Histogram = require('native-hdr-histogram')
-const ProgressBar = require('progress')
 const Client = require('./myhttp')
 const percentiles = require('./percentiles')
 
@@ -11,6 +10,8 @@ function run (opts, cb) {
   cb = cb || noop
 
   const tracker = new EE()
+  tracker.opts = opts
+
   const latencies = new Histogram(1, 10000, 5)
   const requests = new Histogram(1, 1000000, 3)
   const throughput = new Histogram(1, 1000000000, 1)
@@ -25,7 +26,6 @@ function run (opts, cb) {
   const url = URL.parse(opts.url)
 
   opts.pipelining = opts.pipelining || 1
-  opts.progressBarString = opts.progressBarString || 'Autocannon Progress [:bar] :percent'
 
   let counter = 0
   let bytes = 0
@@ -34,20 +34,10 @@ function run (opts, cb) {
   let totalRequests = 0
   let stop = false
   let startTime = Date.now()
-  let progressBar
 
   if (!opts.connections) {
     cb(new Error('connections > 0'))
     return
-  }
-
-  if (opts.renderProgressBar) {
-    progressBar = new ProgressBar(opts.progressBarString, {
-      width: 20,
-      incomplete: ' ',
-      total: opts.duration,
-      clear: true
-    })
   }
 
   // copy over fields so that the client
@@ -93,7 +83,6 @@ function run (opts, cb) {
     counter = 0
     bytes = 0
     tracker.emit('tick')
-    if (progressBar) progressBar.tick()
 
     if (stop) {
       clearInterval(interval)
